requirements.txt
# Core LLM and Agent Framework
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-groq>=0.1.0
langgraph>=0.0.20

# Data Validation and Models
pydantic>=2.0.0

# Google Generative AI
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0

# Optional: For better terminal output
colorama>=0.4.6
